+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `githubinator.enable_context_menu` option to disable/enable access to githubinator in the context menu.
+- Gitlab provider with `githubinator.provider.gitlab.hostname` setting for configuring match.
+
+### Changed
+
+- Github provider hostname is now configured with `githubinator.provider.github.hostname` setting instead of `githubinator.provider.github`.
+
+### Fixed
+
+- Fix regex creation for Github provider so hostname configured in settings is used to match origins.
 
 ## 0.1.0 - 2019-02-28